# mask: don't rely on JSON roundtrip to convert structs